source/ingestion/requirements.txt
functions-framework == 3.*
google-cloud-video-transcoder == 1.*
